requirements.txt
google-genai
python-dotenv
pyyaml
pandas
chromadb
streamlit>=1.40.0
tiktoken
pypdf
python-docx
openpyxl
